Create online_shopping_platform
Compute the total cost of your purchases from an online shopping platform (eg. Shopee or Lazada or Alibaba). You may assume that you may only purchase 1 item in each transaction. The shipping fee will depend on the type of shipping selected. Discounts are available if a voucher code is entered. An optional damage protection (insurance) is offered for any item that costs more than RM250. The damage protection cost is 1% of the cost of the item.  Your program should input the cost of the item, voucher code (if any), shipping option and damage protection (if any). Your program should display (output) the following: - - - The subtotal cost of your item BEFORE shipping fee, discount voucher, and damage protection (if any) The total cost to pay AFTER the discount voucher, shipping fee, and damage protection (if any)) is applied When showing the total cost to pay, be sure to also show the breakdown of the cost that includes discount voucher (if any), shipping fee and damage protection (if any)